Add files via upload
This project is a program written in C language using structures for a hospital that was established in the year 2000. The program takes the input of the name, identity number, admission date of a patient, and the discharge date and calculates the accommodation fee charged at 1000 per day. The program returns the name of the patient and the amount of fee owed to the hospital